refactor the job api to use serverless functions
