+++ .env.local.example
@@ -1,4 +1,4 @@
 CONTENTFUL_SPACE_ID=
 CONTENTFUL_ACCESS_TOKEN=
-
+NEXT_PUBLIC_FUNCTIONS_BASE_URL="http://localhost:3000"
 PEOPLEHR_RSS_FEED_URL=
